+++ graphene-writer/src/main/kotlin/com/graphene/writer/store/key/GrapheneIndexRequest.kt
@@ -4,5 +4,6 @@ import org.elasticsearch.common.xcontent.XContentBuilder
 
 data class GrapheneIndexRequest(
   val id: String,
-  val source: XContentBuilder
+  val source: XContentBuilder,
+  val timestampMillis: Long
 )