+++ graphene-writer/src/main/kotlin/com/graphene/writer/store/key/handler/AbstractElasticsearchKeyStoreHandler.kt
@@ -19,7 +19,6 @@ import org.apache.log4j.Logger
 import org.elasticsearch.action.get.MultiGetRequest
 import org.elasticsearch.client.RequestOptions
 
-// TODO duplicated branch node removing logic
 abstract class AbstractElasticsearchKeyStoreHandler(
   elasticsearchClientFactory: ElasticsearchClientFactory,
   keyStoreHandlerProperty: KeyStoreHandlerProperty
@@ -35,7 +34,7 @@ abstract class AbstractElasticsearchKeyStoreHandler(
   private var templateIndexPattern: String
   private var tenant: String
 
-  private var lastFlushTimeSeconds = DateTimeUtils.currentTimeSeconds()
+  private var lastFlushTimeMillis = DateTimeUtils.currentTimeMillis()
   private var batchSize: Int = 0
   private var flushInterval: Long = 0
 
@@ -54,25 +53,14 @@ abstract class AbstractElasticsearchKeyStoreHandler(
 
     elasticsearchClient = elasticsearchClient(keyStoreHandlerProperty, elasticsearchClientFactory, property)
     elasticsearchClient.createTemplateIfNotExists(templateIndexPattern, templateName(), templateSource())
-    elasticsearchClient.createIndexIfNotExists(index, tenant)
-
-    // 프로퍼티에서 명시한 Index 중 가장 마지막 Offset 을 가져온다.
-    val latestIndex = elasticsearchClient.getLatestIndex(index, tenant)
-    logger.info("latestIndex : $latestIndex")
-//    val aliasDate = indexRollingClient.getAliasDate(latestIndex)
-
-    // 가장 마지막 Offset 의 Index 의 alias 를 확인한다.
+    elasticsearchClient.createIndexIfNotExists(index, tenant, DateTimeUtils.currentTimeMillis(), DateTimeUtils.currentTimeMillis())
 
     keyStoreScheduler = Executors.newSingleThreadScheduledExecutor(NamedThreadFactory(SimpleKeyStoreHandler::class.simpleName!!))
     keyStoreScheduler.scheduleWithFixedDelay(this, 3_000, 500, TimeUnit.MILLISECONDS)
   }
 
   private fun elasticsearchClient(keyStoreHandlerProperty: KeyStoreHandlerProperty, elasticsearchClientFactory: ElasticsearchClientFactory, property: ElasticsearchKeyStoreHandlerProperty): ElasticsearchClient {
-    return if (keyStoreHandlerProperty.rotation.disableRotation()) {
-      elasticsearchClientFactory.createElasticsearchClient(listOf(property.cluster))
-    } else {
-      elasticsearchClientFactory.createIndexRollingEsClient(keyStoreHandlerProperty.rotation, listOf(property.cluster))
-    }
+    return elasticsearchClientFactory.createElasticsearchClient(keyStoreHandlerProperty.rotation, listOf(property.cluster))
   }
 
   override fun handle(grapheneMetric: GrapheneMetric) {
@@ -86,7 +74,7 @@ abstract class AbstractElasticsearchKeyStoreHandler(
         addToBatch(metric)
       }
 
-      if (0 < multiGetRequestContainer.size()) {
+      if (batchSize <= multiGetRequestContainer.size() || lastFlushTimeMillis < DateTimeUtils.currentTimeMillis() - flushInterval) {
         flush()
       }
     } catch (e: Exception) {
@@ -97,14 +85,14 @@ abstract class AbstractElasticsearchKeyStoreHandler(
 
   private fun addToBatch(metric: GrapheneMetric) {
     multiGetRequestContainer.add(type, metric)
-
-    if (batchSize <= multiGetRequestContainer.size() || lastFlushTimeSeconds < DateTimeUtils.currentTimeSeconds() - flushInterval) {
-      flush()
-    }
   }
 
   private fun flush() {
-    elasticsearchClient.createIndexIfNotExists(index, tenant)
+    if (0 >= multiGetRequestContainer.size()) {
+      return
+    }
+
+    elasticsearchClient.createIndexIfNotExists(index, tenant, multiGetRequestContainer.from, multiGetRequestContainer.to)
 
     val multiGetResponse = elasticsearchClient.mget(multiGetRequestContainer.multiGetRequest, RequestOptions.DEFAULT)
     val bulkRequest = mutableListOf<GrapheneIndexRequest>()
@@ -119,7 +107,7 @@ abstract class AbstractElasticsearchKeyStoreHandler(
         continue
       }
 
-      val metric = multiGetRequestContainer.metrics[response.id]
+      val metric = multiGetRequestContainer.metrics["${response.index}_${response.id}"]
       bulkRequest.addAll(mapToGrapheneIndexRequests(metric))
     }
 
@@ -132,17 +120,37 @@ abstract class AbstractElasticsearchKeyStoreHandler(
       logger.error("Fail to index metric key, reason : ${bulkResponse.buildFailureMessage()}")
     }
     multiGetRequestContainer = MultiGetRequestContainer()
-    lastFlushTimeSeconds = DateTimeUtils.currentTimeSeconds()
+    lastFlushTimeMillis = DateTimeUtils.currentTimeMillis()
   }
 
   private inner class MultiGetRequestContainer(
     val multiGetRequest: MultiGetRequest = MultiGetRequest(),
-    val metrics: MutableMap<String, GrapheneMetric> = mutableMapOf()
+    val metrics: MutableMap<Index, GrapheneMetric> = mutableMapOf(),
+    var from: Long? = null,
+    var to: Long? = null
   ) {
 
     fun add(type: String, metric: GrapheneMetric) {
-      metrics[metric.getId()] = metric
-      multiGetRequest.add(MultiGetRequest.Item(elasticsearchClient.getIndexWithDate(index, tenant), type, metric.getId()))
+      val timestampMillis = metric.timestampMillis()
+      val indexWithDate = elasticsearchClient.getIndexWithDate(index, tenant, timestampMillis)
+
+      if (!metrics.contains("${indexWithDate}_${metric.getId()}")) {
+        metrics["${indexWithDate}_${metric.getId()}"] = metric
+        multiGetRequest.add(MultiGetRequest.Item(indexWithDate, type, metric.getId()))
+
+        if (Objects.isNull(from) && Objects.isNull(to)) {
+          from = timestampMillis
+          to = timestampMillis
+        }
+
+        if (timestampMillis < from!!) {
+          from = timestampMillis
+        }
+
+        if (to!! < timestampMillis) {
+          to = timestampMillis
+        }
+      }
     }
 
     fun size(): Int {
@@ -165,3 +173,5 @@ abstract class AbstractElasticsearchKeyStoreHandler(
 
   abstract fun templateName(): String
 }
+
+typealias Index = String