+++ graphene-writer/src/main/kotlin/com/graphene/writer/store/key/handler/SimpleKeyStoreHandler.kt
@@ -35,7 +35,7 @@ class SimpleKeyStoreHandler(
       try {
         val graphiteKeyPart = graphiteKeySb.toString()
         val id = "${metric.getTenant()}_$graphiteKeyPart"
-        grapheneIndexRequests.add(GrapheneIndexRequest(id, source(metric.getTenant(), graphiteKeyPart, depth, isLeaf(depth, parts))))
+        grapheneIndexRequests.add(GrapheneIndexRequest(id, source(metric.getTenant(), graphiteKeyPart, depth, isLeaf(depth, parts)), metric.timestampMillis()))
       } catch (e: Exception) {
         throw IllegalStateException("Invokes illegal state in map to index requests", e)
       }
@@ -71,7 +71,8 @@ class SimpleKeyStoreHandler(
       {
         "settings": {
           "number_of_replicas": 0,
-          "number_of_shards": 5
+          "number_of_shards": 5,
+          "auto_expand_replicas": "0-1"
         },
         "mappings": {
           "path": {